fixed user layout re-renders & refactor
