+++ pages/user/calendar/workout/[el].tsx
@@ -4,11 +4,16 @@ import { useRouter } from "next/router"
 import CalendarContainer from "@/components/Calendar/calendarContainer"
 import HeadLinkBack from "@/components/HeadLinkBack/headLinkBack"
 import ItemWorkout from "@/components/ItemWorkout/ItemWorkout"
-import { LayoutUser } from "@/containers/Layout-user/layoutUser"
-import { WorkoutUl, ListWorkoutName, ItemWorkoutDiv } from "@/components/ItemWorkout/styledWorkout"
-import { workoutDay, IWorkoutDayType } from "@/models/workout/workoutDay"
 
-export default LayoutUser(function () {
+import {
+  WorkoutUl,
+  ListWorkoutName,
+  ItemWorkoutDiv,
+} from "@/components/ItemWorkout/styledWorkout"
+import { workoutDay, IWorkoutDayType } from "@/models/workout/workoutDay"
+import { WithLayout } from "@/containers/Layout-user/withLayout"
+
+const WorkoutItem = () => {
   const { asPath } = useRouter()
   const [arrWorkoutDay, setWorkoutDay] = useState<IWorkoutDayType[]>(workoutDay)
 
@@ -58,5 +63,6 @@ export default LayoutUser(function () {
       </div>
     </CalendarContainer>
   )
-})
+}
 
+export default WithLayout(WorkoutItem)
